Fixed UI tests for packages

+++ Tests/FeatherWidgets.TestUI.Arrangements/Packaging/StaticContent/Libraries/ImportEditedVideosStructure.cs
@@ -52,7 +52,10 @@ namespace FeatherWidgets.TestUI.Arrangements
         {            
             ServerOperations.Videos().DeleteLibrary(AlbumName, false);
             ServerOperations.Pages().DeleteAllPages();
-            ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
+            if (System.IO.Directory.Exists(this.tempFolderPath))
+                ServerOperations.ModuleBuilder().DeleteDirectory(this.tempFolderPath);
+
             ServerOperations.Packaging().DeleteAllPackagesFromDB();
 
             for (int i = 0; i < this.widgetTemplatesNames.Length; i++)
